fix: Removed the lint dependencies and config to fix a compile issue
There seems to be a problem with one of the dependencies I installed
for linting - they lint fine, the linted code runs fine without
the dependencies, but once they are installed, there is a strange error.
I assume this is a stability issue with one of them because the code looks
fine and runs fine without them.
I have also left them intact in branch "janis-zalitis-myreads-linted"
to investigate later.